[IR] refine the ir folder in fluid. (#54292)
