+++ .gitignore
@@ -96,4 +96,4 @@ paddle/phi/api/profiler/__init__.py
 python/paddle/incubate/fleet/parameter_server/pslib/ps_pb2.py
 paddle/phi/kernels/fusion/cutlass/conv2d/generated/*
 python/paddle/fluid/incubate/fleet/parameter_server/pslib/ps_pb2.py
-paddle/fluid/translator/op_compat_info.cc
+paddle/fluid/ir_adaptor/translator/op_compat_info.cc
