+++ paddle/fluid/CMakeLists.txt
@@ -9,8 +9,8 @@ add_subdirectory(eager)
 add_subdirectory(prim)
 add_subdirectory(jit)
 if(WITH_NEWIR)
-  add_subdirectory(dialect)
-  add_subdirectory(translator)
+  add_subdirectory(ir)
+  add_subdirectory(ir_adaptor)
 endif()
 # NOTE: please add subdirectory inference at last.
 add_subdirectory(inference)
