+++ paddle/fluid/ir/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(interface)
+add_subdirectory(dialect)
