+++ paddle/fluid/ir/dialect/pd_attribute.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/dialect/pd_attribute.h"
+#include "paddle/fluid/ir/dialect/pd_attribute.h"
 
 namespace paddle {
 namespace dialect {
