+++ paddle/fluid/ir/dialect/pd_dialect.cc
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/dialect/pd_dialect.h"
-#include "paddle/fluid/dialect/pd_attribute.h"
+#include "paddle/fluid/ir/dialect/pd_dialect.h"
+#include "paddle/fluid/ir/dialect/pd_attribute.h"
 // NOTE(zhangbo9674): File pd_op.h is generated by op_gen.py, see details in
-// paddle/fluid/dialect/CMakeLists.txt.
-#include "paddle/fluid/dialect/pd_op.h"
-#include "paddle/fluid/dialect/pd_type.h"
-#include "paddle/fluid/dialect/pd_type_storage.h"
-#include "paddle/fluid/dialect/utils.h"
+// paddle/fluid/ir/dialect/CMakeLists.txt.
 #include "paddle/fluid/framework/convert_utils.h"
 #include "paddle/fluid/framework/data_type.h"
+#include "paddle/fluid/ir/dialect/pd_op.h"
+#include "paddle/fluid/ir/dialect/pd_type.h"
+#include "paddle/fluid/ir/dialect/pd_type_storage.h"
+#include "paddle/fluid/ir/dialect/utils.h"
 #include "paddle/ir/core/dialect_interface.h"
 #include "paddle/phi/core/dense_tensor.h"
 
@@ -99,10 +99,10 @@ void PaddleDialect::initialize() {
 
   // NOTE(zhangbo9674): GET_OP_LIST is defined in pd_op.h which is
   // generated by op_gen.py, see details in
-  // paddle/fluid/dialect/CMakeLists.txt.
+  // paddle/fluid/ir/dialect/CMakeLists.txt.
   RegisterOps<
 #define GET_OP_LIST
-#include "paddle/fluid/dialect/pd_op.h"  // NOLINT
+#include "paddle/fluid/ir/dialect/pd_op.h"  // NOLINT
       >();
 
   RegisterInterfaces<ParameterConvertInterface>();
