+++ paddle/fluid/ir/dialect/pd_interface.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "paddle/fluid/dialect/utils.h"
+#include "paddle/fluid/ir/dialect/utils.h"
 #include "paddle/ir/core/op_base.h"
 
 using OpInfoTuple = std::tuple<std::vector<paddle::dialect::OpInputInfo>,
