+++ paddle/fluid/ir_adaptor/translator/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(PD_PROGRAM_TRANSLATOR_SOURCE_DIR "${CMAKE_CURRENT_LIST_DIR}")
 set(PD_PROGRAM_TRANSLATOR_BINARY_DIR
-    "${PADDLE_BINARY_DIR}/paddle/fluid/translator")
+    "${PADDLE_BINARY_DIR}/paddle/fluid/ir_adaptor/translator/")
 
 set(op_gen_file ${PD_PROGRAM_TRANSLATOR_SOURCE_DIR}/op_compat_gen.py)
 set(op_compat_yaml_file ${PADDLE_SOURCE_DIR}/paddle/phi/api/yaml/op_compat.yaml)
