+++ paddle/fluid/ir_adaptor/translator/op_compat_info.cc.j2
@@ -1,4 +1,4 @@
-#include "paddle/fluid/translator/op_compat_info.h"
+#include "paddle/fluid/ir_adaptor/translator/op_compat_info.h"
 
 namespace paddle {
 namespace translator {
