+++ paddle/fluid/ir_adaptor/translator/program_translator.cc
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/translator/program_translator.h"
+#include "paddle/fluid/ir_adaptor/translator/program_translator.h"
 
 #include <unordered_map>
 
 #include "glog/logging.h"
 
 #include "paddle/fluid/framework/program_desc.h"
-#include "paddle/fluid/translator/op_translator.h"
-#include "paddle/fluid/translator/type_translator.h"
+#include "paddle/fluid/ir_adaptor/translator/op_translator.h"
+#include "paddle/fluid/ir_adaptor/translator/type_translator.h"
 #include "paddle/ir/core/attribute.h"
 #include "paddle/ir/core/block.h"
 #include "paddle/ir/core/builtin_op.h"
