+++ paddle/fluid/ir_adaptor/translator/translate.h
@@ -14,30 +14,15 @@
 
 #pragma once
 
-#include "paddle/fluid/dialect/pd_type_storage.h"
-#include "paddle/ir/core/type.h"
+#include <memory>
+
+#include "paddle/fluid/framework/program_desc.h"
+#include "paddle/fluid/ir/dialect/pd_dialect.h"
+#include "paddle/ir/core/program.h"
 
 namespace paddle {
-namespace dialect {
-///
-/// \brief Define built-in parametric types.
-///
-class DenseTensorType : public ir::Type {
- public:
-  using Type::Type;
 
-  DECLARE_TYPE_UTILITY_FUNCTOR(DenseTensorType, DenseTensorTypeStorage);
+std::unique_ptr<::ir::Program> TranslateLegacyProgramToProgram(
+    const ::paddle::framework::ProgramDesc& legacy_program);
 
-  const ir::Type &dtype() const;
-
-  const phi::DDim &dims() const;
-
-  const phi::DataLayout &data_layout() const;
-
-  const phi::LoD &lod() const;
-
-  const size_t &offset() const;
-};
-
-}  // namespace dialect
 }  // namespace paddle
