+++ paddle/fluid/ir_adaptor/translator/type_translator.h
@@ -19,7 +19,7 @@
 
 #include "paddle/fluid/framework/op_desc.h"
 #include "paddle/fluid/framework/var_desc.h"
-#include "paddle/fluid/translator/program_translator.h"
+#include "paddle/fluid/ir_adaptor/translator/program_translator.h"
 #include "paddle/ir/core/builtin_type.h"
 #include "paddle/ir/core/dialect.h"
 #include "paddle/ir/core/ir_context.h"
