+++ test/cpp/ir/core/ir_program_test.cc
@@ -14,10 +14,10 @@
 
 #include <gtest/gtest.h>
 
-#include "paddle/fluid/dialect/pd_dialect.h"
-#include "paddle/fluid/dialect/pd_interface.h"
-#include "paddle/fluid/dialect/pd_type.h"
-#include "paddle/fluid/dialect/utils.h"
+#include "paddle/fluid/ir/dialect/pd_dialect.h"
+#include "paddle/fluid/ir/dialect/pd_interface.h"
+#include "paddle/fluid/ir/dialect/pd_type.h"
+#include "paddle/fluid/ir/dialect/utils.h"
 #include "paddle/ir/core/block.h"
 #include "paddle/ir/core/builtin_attribute.h"
 #include "paddle/ir/core/builtin_dialect.h"
